Fixed an issue with changing configuration of Jetty HTTP server #server-824

+++ src/main/java/tigase/http/jetty/JettyStandaloneHttpServer.java
@@ -139,9 +139,9 @@ public class JettyStandaloneHttpServer
 	protected void unregisterConnector(ServerConnector connector, boolean secure) {
 		server.removeConnector(connector);
 		if (secure) {
-			httpsPorts.remove(connector.getPort());
+			httpsPorts.remove((Integer) connector.getPort());
 		} else {
-			httpPorts.remove(connector.getPort());
+			httpPorts.remove((Integer) connector.getPort());
 		}
 	}
